[#22] Refactor folder organization for indoor directions components
