+++ components/buildings/buildingView/buildingNoFloors/index.js
