+++ components/buildings/buildingView/buildingNoFloors/styles.js
