+++ components/buildings/buildingView/buildingWithFloors/styles.js
