+++ components/buildings/buildingView/index.js
