+++ components/buildings/floorPlans/H/Hall9.js
