+++ components/buildings/index.js
@@ -1,0 +1,5 @@
+/* istanbul ignore file */
+export { default as Hall8 } from './floorPlans/H/Hall8';
+export { default as Hall9 } from './floorPlans/H/Hall9';
+export { default as Vl1 } from './floorPlans/VL/Vl1';
+export { default as Vl2 } from './floorPlans/VL/Vl2';
