+++ components/home/location/index.js
@@ -4,9 +4,9 @@ import React, { Component } from 'react';
 import { View, Image, Text } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import getCurrentLocation from './LocationServices';
-import buildings from '../../assets/polygons/polygons';
+import buildings from '../../../assets/polygons/polygons';
 import styles from './styles';
-import locateMe from '../../assets/icons/locate-me.png';
+import locateMe from '../../../assets/icons/locate-me.png';
 
 export default class Location extends Component {
   constructor(props) {
