+++ components/home/location/styles.js
