--- components/indoorPlans/index.js
@@ -1,5 +1,0 @@
-/* istanbul ignore file */
-export { default as Hall8 } from './buildingComponents/H/Hall8';
-export { default as Hall9 } from './buildingComponents/H/Hall9';
-export { default as Vl1 } from './buildingComponents/VL/Vl1';
-export { default as Vl2 } from './buildingComponents/VL/Vl2';
